utils/asymmetricKeys.py
# utils/asymmetricKeys.py

from cryptography.hazmat.primitives.asymmetric import rsa
from cryptography.hazmat.primitives import serialization

_private_key = rsa.generate_private_key(public_exponent=65537, key_size=2048)

_public_key = _private_key.public_key()

def get_private_key():
    return _private_key

def get_public_key_pem():
    return _public_key.public_bytes(
        encoding=serialization.Encoding.PEM,
        format=serialization.PublicFormat.SubjectPublicKeyInfo
    ).decode()
